+++ imap_processing/codice/utils.py
@@ -15,6 +15,7 @@ import numpy as np
 import xarray as xr
 
 from imap_processing.codice import constants
+from imap_processing.codice.constants import CODICEAPID
 
 
 @dataclass
@@ -44,45 +45,6 @@ class ViewTabInfo:
     sensor: int
     three_d_collapsed: int
     view_id: int
-
-
-class CODICEAPID(IntEnum):
-    """Create ENUM for CoDICE APIDs."""
-
-    COD_AUT = 1120
-    COD_BOOT_HK = 1121
-    COD_BOOT_MEMDMP = 1122
-    COD_COUNTS_COMMON = 1135
-    COD_NHK = 1136
-    COD_EVTMSG = 1137
-    COD_MEMDMP = 1138
-    COD_SHK = 1139
-    COD_RTS = 1141
-    COD_DIAG_CDHFPGA = 1144
-    COD_DIAG_SNSR_HV = 1145
-    COD_DIAG_OPTC_HV = 1146
-    COD_DIAG_APDFPGA = 1147
-    COD_DIAG_SSDFPGA = 1148
-    COD_DIAG_FSW = 1149
-    COD_DIAG_SYSVARS = 1150
-    COD_LO_IAL = 1152
-    COD_LO_PHA = 1153
-    COD_LO_SW_PRIORITY_COUNTS = 1155
-    COD_LO_SW_SPECIES_COUNTS = 1156
-    COD_LO_NSW_SPECIES_COUNTS = 1157
-    COD_LO_SW_ANGULAR_COUNTS = 1158
-    COD_LO_NSW_ANGULAR_COUNTS = 1159
-    COD_LO_NSW_PRIORITY_COUNTS = 1160
-    COD_LO_INST_COUNTS_AGGREGATED = 1161
-    COD_LO_INST_COUNTS_SINGLES = 1162
-    COD_HI_IAL = 1168
-    COD_HI_PHA = 1169
-    COD_HI_INST_COUNTS_AGGREGATED = 1170
-    COD_HI_INST_COUNTS_SINGLES = 1171
-    COD_HI_OMNI_SPECIES_COUNTS = 1172
-    COD_HI_SECT_SPECIES_COUNTS = 1173
-    COD_HI_INST_COUNTS_PRIORITIES = 1174
-    COD_CSTOL_CONFIG = 2457
 
 
 class CoDICECompression(IntEnum):
