+++ imap_processing/tests/codice/test_codice_l1a.py
@@ -9,20 +9,18 @@ caused too much complexity.
 """
 
 import logging
-from pathlib import Path
 from unittest.mock import patch
 
 import cdflib
 import numpy as np
 import pytest
-import xarray as xr
 from imap_data_access import ProcessingInputCollection
 
 from imap_processing import imap_module_directory
 from imap_processing.cdf.utils import load_cdf, write_cdf
 from imap_processing.codice.codice_l1a import process_l1a
 from imap_processing.codice.codice_l1a_de import l1a_direct_event
-from imap_processing.codice.utils import CODICEAPID, process_by_table_id
+from imap_processing.codice.utils import CODICEAPID
 from imap_processing.tests.codice.conftest import (
     VALIDATION_FILE_DATE,
     VALIDATION_FILE_VERSION,
@@ -36,40 +34,6 @@ pytestmark = pytest.mark.external_test_data
 # and num_spins max = 16 in the current CoDICE timing model. That yields a
 # worst-case delta of 128 s = 128000000000 ns.
 EXPECTED_EPOCH_DELTA_VALIDMAX = 128000000000
-
-
-def test_process_by_table_id_keeps_voltage_table_1d():
-    """Shared support data should not be expanded across epoch groups."""
-
-    unpacked_dataset = xr.Dataset(
-        data_vars={
-            "view_id": ("epoch", np.array([0, 0], dtype=np.int16)),
-            "pkt_apid": ("epoch", np.array([1156, 1156], dtype=np.int16)),
-            "plan_id": ("epoch", np.array([7, 7], dtype=np.int16)),
-            "plan_step": ("epoch", np.array([3, 3], dtype=np.int16)),
-            "table_id": ("epoch", np.array([1, 2], dtype=np.int16)),
-            "sample": ("epoch", np.array([11, 22], dtype=np.int16)),
-        },
-        coords={"epoch": np.array([20, 10], dtype=np.int64)},
-    )
-
-    voltage_table = np.array([1.0, 2.0, 3.0], dtype=np.float64)
-
-    def _process_fn(group_ds, lut_file, table_id, view_id, apid, plan_id, plan_step):
-        # Mock processed data with non-epoch data variables
-        return xr.Dataset(
-            data_vars={
-                "sample": ("epoch", group_ds["sample"].values),
-                "voltage_table": ("esa_step", voltage_table),
-            },
-            coords={"epoch": group_ds["epoch"].values, "esa_step": np.arange(3)},
-        )
-
-    processed = process_by_table_id(unpacked_dataset, Path("unused.json"), _process_fn)
-
-    assert processed["epoch"].values.tolist() == [10, 20]
-    assert processed["voltage_table"].dims == ("esa_step",)
-    np.testing.assert_array_equal(processed["voltage_table"].values, voltage_table)
 
 
 def assert_epoch_delta_cdf_metadata(cdf_file):
